requirements.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
confusable-homoglyphs==3.2.0
dj-database-url==0.5.0
Django==3.1.2
django-debug-toolbar==3.1.1
django-registration==3.1.1
gunicorn==20.0.4
idna==2.10
psycopg2-binary==2.8.6
python-decouple==3.3
python-http-client==3.3.1
pytz==2020.1
requests==2.24.0
sendgrid==6.4.7
sqlparse==0.4.1
starkbank-ecdsa==1.1.0
urllib3==1.25.10
whitenoise==5.2.0
